reading params and config file

+++ Makefile
@@ -1,8 +1,14 @@
-compile: dir-create
-	gcc npt.c -o bin/npt
+compile: clean dir-create main.o
+	gcc main.o utils.o -o bin/npt
 
 dir-create:
 	mkdir -p bin
 
+main.o: utils.o
+	gcc -o main.o main.c -c
+
+utils.o:
+	gcc -o utils.o utils.c -c
+
 clean:
-	rm -rf bin/ *.o
+	clear && rm -rf bin/ *.o
